requirements.txt
Django==6.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
gunicorn==21.2.0
dj-database-url==3.1.0
python-dotenv==1.0.0
pyTelegramBotAPI==4.11.0
whitenoise==6.11.0